% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computing-accuracy.R
\name{compute_mae_and_isp}
\alias{compute_mae_and_isp}
\title{Calculate MAE and ISP}
\usage{
compute_mae_and_isp(file1, file2)
}
\arguments{
\item{file1}{file path for first grades csv}

\item{file2}{file path for second grades csv}
}
\value{
a list
}
\description{
This function calculates the proportion of identical scores
and mean absolute error of rubric items
between two Gradescope evaluation csv
}
