% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-metadata.R
\name{create_metadata_json}
\alias{create_metadata_json}
\title{Create Metadata JSON}
\usage{
create_metadata_json(
  dir = "",
  department,
  course_number,
  course_name,
  upper_div,
  year,
  semester,
  assignment_name,
  question_number,
  question_name = "",
  mode_of_question,
  medium_of_answer,
  content_of_answer
)
}
\arguments{
\item{dir}{directory where metadata.json should be saved}

\item{department}{department abbreviation (e.g. "STAT", "DATA")}

\item{course_number}{course number}

\item{course_name}{name of course (e.g. "Introduction to Probability")}

\item{upper_div}{if the course is upper-division or lower-division}

\item{year}{year of the course}

\item{semester}{semester of the course (e.g. "Spring", "Fall")}

\item{assignment_name}{name of assignment (e.g. "Midterm 1")}

\item{question_number}{number of question, including subparts (e.g. "5c")}

\item{question_name}{optionally name of question (e.g. "Data Visualization")}

\item{mode_of_question}{mode of question (e.g. "fill in the blank", "open-ended")}

\item{medium_of_answer}{medium of student's answer (e.g. "handwritten", "typed")}

\item{content_of_answer}{expected content of student's answer (e.g. "English", "math", "code")}
}
\description{
This function creates a JSON file that stores metadata
about the assignment and its course. The contents of the saved JSON
file is also printed out.
}
