% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deidentify-grades.R
\name{deidentify_gradescope_evals}
\alias{deidentify_gradescope_evals}
\title{Deidentify Gradescope Evaluations}
\usage{
deidentify_gradescope_evals(
  gs_csv_path,
  ids_csv_path,
  output_path,
  ignored_nrows = 3
)
}
\arguments{
\item{gs_csv_path}{path to exported Gradescope csv}

\item{ids_csv_path}{path to de-identified table csv}

\item{output_path}{path to de-idenfitied Gradescope output}

\item{ignored_nrows}{how many lines at the end of csv to exclude}
}
\description{
This function de-identifies the exported
Gradescope csv by swapping all columns with
identifiable data with the de-identified IDs
in an existing table. Note that Gradescope csv
must have a column called \verb{Assignment Submission ID} and the ids table
must have two columns: \code{SID}, \code{Generated_ID}
}
