% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-expert-updates.R
\name{expert_updates_table}
\alias{expert_updates_table}
\title{Calculate Differences in Expert Grades After QA/QC}
\usage{
expert_updates_table(pre_dir, post_dir)
}
\arguments{
\item{pre_dir}{directory before QA/QC}

\item{post_dir}{directory after QA/QC}
}
\value{
dataframe
}
\description{
Find the differences in expert grades after comparing to AI grades
and to student grades.
}
