% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computing-accuracy.R
\name{export_grading_differences_gt}
\alias{export_grading_differences_gt}
\title{Export Differences between Two Graders as GT Table}
\usage{
export_grading_differences_gt(find_differences)
}
\arguments{
\item{find_differences}{object from the function find_differences_table()}
}
\value{
a gt object
}
\description{
This function exports the differences in rubric items
in GT format between two graders with mismatched rubrics highlighted.
}
