% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computing-accuracy.R
\name{find_differences_gt}
\alias{find_differences_gt}
\title{Find Differences as GT Table}
\usage{
find_differences_gt(file1, file2)
}
\arguments{
\item{file1}{file path for first grades for comparison}

\item{file2}{file path for second grades for comparison}
}
\value{
a gt object
}
\description{
This function displays a table of differences in rubric items
in GT format, with mismatched rubrics highlighted.
}
