% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computing-accuracy.R
\name{find_differences_wrt_experts}
\alias{find_differences_wrt_experts}
\title{Find Differences with respect to Experts Table}
\usage{
find_differences_wrt_experts(experts_file, ai_file)
}
\arguments{
\item{experts_file}{file with expert graders}

\item{ai_file}{file with AI graders}
}
\value{
a list of a df and a matrix
}
\description{
Find the differences between AI grading and experts and return
all graded assignments that are different and a matrix of which rubrics
are mismatched.
}
