% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computing-accuracy.R
\name{find_differences_wrt_experts_xlsx}
\alias{find_differences_wrt_experts_xlsx}
\title{Find Differences with respect to Experts as Excel Spreadsheet}
\usage{
find_differences_wrt_experts_xlsx(
  file1,
  file2,
  existing,
  sheet_name,
  dir = "."
)
}
\arguments{
\item{file1}{file path for first grades for comparison}

\item{file2}{file path for second grades for comparison}

\item{existing}{if workbook exists}

\item{sheet_name}{name of sheet in workbook}

\item{dir}{optionally, where workbook is saved}
}
\description{
This function saves an Excel spreadsheet with the differences
in rubric items with respect to experts with mismatched rubrics highlighted.
}
