% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computing-accuracy.R
\name{find_differences_wrt_students}
\alias{find_differences_wrt_students}
\title{Find Differences with respect to Students Table}
\usage{
find_differences_wrt_students(experts_file, student_file, ai_file, ai_diffs)
}
\arguments{
\item{experts_file}{file with experts graders}

\item{student_file}{file with student graders}

\item{ai_file}{file with ai graders}

\item{ai_diffs}{find_differences results for experts v. AI}
}
\value{
a list of a df and a matrix
}
\description{
Find the differences between students grading and experts and return
all graded assignments that are different and a matrix of which rubrics
are mismatched. Note that all mismatches between AI and expert grading
are removed to prevent redundancy.
}
