% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-results-table.R
\name{generate_gt_results_table}
\alias{generate_gt_results_table}
\title{Generate Results Table in GT Format}
\usage{
generate_gt_results_table(results_table)
}
\arguments{
\item{results_table}{results_table as generated by \code{generate_results_table}}
}
\value{
a gt object
}
\description{
This function displays a results table in GT format.
}
