% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-results-table.R
\name{generate_metadata_table}
\alias{generate_metadata_table}
\title{Generate Metadata Table}
\usage{
generate_metadata_table(dir = ".")
}
\arguments{
\item{dir}{optionally, specify directory}
}
\value{
a dataframe
}
\description{
This function parses through the data/scores folders and generates
the metadata table using the metadata.jsons for each assignment.
}
