% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deidentify-grades.R
\name{generate_rubric_texts}
\alias{generate_rubric_texts}
\title{Generate Rubric Texts}
\usage{
generate_rubric_texts(
  csv_path,
  output_folder,
  ignored_nrows = 3,
  existing = TRUE,
  pensieve = FALSE
)
}
\arguments{
\item{csv_path}{path to exported Gradescope evaluations csv}

\item{output_folder}{folder for rubric_items.csv}

\item{ignored_nrows}{how many of the last lines to ignore}

\item{existing}{if there is an existing rubric_items.csv}

\item{pensieve}{if source is exported from Pensieve}
}
\value{
a dataframe with rubric items in "R1", "R2",... format
}
\description{
This function generates or updated the rubric_items.csv
with a grades csv file's rubric items. The grades csv is
saved with "R1", "R2", ... labels for the rubric items
as a new csv, and this function returns this new grades csv
as a dataframe.
}
