% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computing-accuracy.R
\name{identical_score_prop}
\alias{identical_score_prop}
\title{Calculate Proportion of Identical Scores}
\usage{
identical_score_prop(eval1, eval2)
}
\arguments{
\item{eval1}{first dataframe of Gradescope evaluations}

\item{eval2}{first dataframe of Gradescope evaluations}
}
\value{
a single proportion
}
\description{
This function calculates the proportion of identical scores
between two Gradescope evaluation dataframes
}
