% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computing-accuracy.R
\name{normalize_full_credit}
\alias{normalize_full_credit}
\title{Normalize Full Credit}
\usage{
normalize_full_credit(evals)
}
\arguments{
\item{evals}{dataframe of Gradescope evaluations}
}
\description{
This function "normalizes" the "Full Credit" toggle
on the evaluations by making all other rubric items
TRUE to allow for comparisons across individual rubric
items.
}
