% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge-grades.R
\name{read_evals}
\alias{read_evals}
\title{Read Gradescope Evaluations CSV}
\usage{
read_evals(csv_path, ignore_nrows = 5)
}
\arguments{
\item{csv_path}{path to exported Gradescope evaluations csv}

\item{ignore_nrows}{how many of the last lines to ignore}
}
\description{
This function reads in the .csv file from Gradescope's
"Export Evaluations" functionality. The formatting of this
file is not in tidy format because the last few lines include
Point Values, Rubric Numbers and Scoring Method. These are
excluded for the sake of this package.
}
