% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-metadata.R
\name{update_scores}
\alias{update_scores}
\title{Update student scores based on metadata rubric items}
\usage{
update_scores(csv, metadata = "./metadata.json", overwrite = TRUE, calibrated)
}
\arguments{
\item{csv}{the path to the student CSV file.}

\item{metadata}{the path to the metadata JSON file}

\item{overwrite}{whether to overwrite the original
CSV file with updated scores}

\item{calibrated}{which rubric to apply}
}
\value{
A data frame containing the original student data with an
updated \code{Score} column.
}
\description{
Computes total scores for each student submission by multiplying
rubric item responses by their corresponding point values defined
in a metadata JSON file.
}
