% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-metadata.R
\name{update_scores_in_metadata}
\alias{update_scores_in_metadata}
\title{Update Scores in metadata JSON}
\usage{
update_scores_in_metadata(folder = "./", file = "./metadata.json")
}
\arguments{
\item{folder}{the directory containing the CSV files}

\item{file}{the path to the metadata JSON file}
}
\description{
Reads student and expert grading CSV files, computes
\code{n_submissions} and \code{mean_score}, and updates the corresponding fields in a
metadata JSON file.
}
