% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deidentify-gradescope.R
\name{deidentify_graders}
\alias{deidentify_graders}
\title{Deidentify Graders}
\usage{
deidentify_graders(csv_path, roster_csv, output_path)
}
\arguments{
\item{csv_path}{path to exported Gradescope evaluations csv}

\item{roster_csv}{path to roster csv}

\item{output_path}{path to save deidentified grades}
}
\description{
This function swaps the original grader names with fake names
from a roster and saves the deidentified grades as a csv
and the lookup table for the graders as a csv.
}
