% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computing-accuracy.R
\name{export_grading_differences_xlsx}
\alias{export_grading_differences_xlsx}
\title{Export Differences between Two Graders as Excel Spreadsheet}
\usage{
export_grading_differences_xlsx(
  find_differences,
  existing,
  sheet_name,
  dir = "."
)
}
\arguments{
\item{find_differences}{object from the function find_differences_table()}

\item{existing}{if workbook exists}

\item{sheet_name}{name of sheet in workbook}

\item{dir}{optionally, where workbook is saved}
}
\description{
This function saves an Excel spreadsheet with the differences
in rubric items between two graders with mismatched rubrics highlighted.
}
