% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-results-table.R
\name{generate_results_table}
\alias{generate_results_table}
\title{Generate Results Table}
\usage{
generate_results_table(dir = ".")
}
\arguments{
\item{dir}{optionally, specify directory}
}
\value{
a dataframe
}
\description{
This function parses through the data/scores folders and generates
the results table for grading accuracy comparison.
}
