% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computing-accuracy.R
\name{isp}
\alias{isp}
\title{Calculate Proportion of Identical Scores}
\usage{
isp(eval1, eval2, rubric_matching_list)
}
\arguments{
\item{eval1}{first dataframe of Gradescope evaluations}

\item{eval2}{second dataframe of Gradescope evaluations}

\item{rubric_matching_list}{this is only to check if the comparison is valid}
}
\value{
a single proportion
}
\description{
This function calculates the proportion of identical scores
between two Gradescope evaluation dataframes
}
