% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computing-accuracy.R
\name{normalize_full_credit}
\alias{normalize_full_credit}
\title{Normalize Full Credit}
\usage{
normalize_full_credit(evals, full_credit, rubric_items)
}
\arguments{
\item{evals}{dataframe of Gradescope evaluations}

\item{full_credit}{column index or name for full credit column}

\item{rubric_items}{vector of rubric items that sum up to full credit}
}
\value{
normalized evals dataframe
}
\description{
This function "normalizes" the "Full Credit" toggle
on the evaluations by making all other rubric items
TRUE to allow for comparisons across individual rubric
items. Note that \code{rubric_items} can be a string vector of
column names or a numeric vector of column indices.
}
