% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computing-accuracy.R
\name{rubric_mean_absolute_error}
\alias{rubric_mean_absolute_error}
\title{Mean Absolute Error of Rubric Items}
\usage{
rubric_mean_absolute_error(eval1, eval2, rubric_matching_list)
}
\arguments{
\item{eval1}{first dataframe of Gradescope evaluations}

\item{eval2}{second dataframe of Gradescope evaluations}

\item{rubric_matching_list}{vector of rubric items that sum up to full credit}
}
\value{
normalized evals dataframe
}
\description{
If we had a rubric with two items, both worth one point, the error contribution of one submission would be:\cr
For expert answer \{1, 0\}, \cr
0 if grader is \{1, 0\} \cr
1 if grader is \{0, 0\} or \{1, 1\} \cr
2 if grader is \{0, 1\} \cr
It's recommended to \code{normalize_full_credit()} for \code{eval1} and \code{eval2} prior
to using this function.
}
