% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-metadata.R
\name{validate_metadata_json}
\alias{validate_metadata_json}
\title{Validate Metadata JSON}
\usage{
validate_metadata_json(file = "./metadata.json", verbose = F)
}
\arguments{
\item{file}{file path where metadata.json should be saved}

\item{verbose}{if course information should be printed out}
}
\description{
This function validates a JSON file that stores metadata
about the assignment, its course and rubric items. The contents of the saved JSON
file can be printed out.
}
